+++ Models/EntityStatus.cs
@@ -1,0 +1,11 @@
+namespace Review.Api.Models;
+
+public enum EntityStatus
+{
+    Updated,
+    NoChanges,
+    Deleted,
+    NotFound,
+    Forbidden,
+    InvalidReference
+}